==> code/audio_features/1.scraping_audio_features_dr_sample_tracks.ipynb <==
# Getting data 3 - audio features of sample dr artist album tracks

Now that I got the sample of dr tracks, I need to get the audio features of said tracks.

I got a csv with the following data:
- track id
- track artist
- track title
- track explicity

<br/>

Next steps are:
- getting audio features from dr tracks

# importing needed libraries

import pandas as pd
import numpy as np
import getpass
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

# importing data csv

dr_sample_tracks = pd.read_csv('../data/dr_tracks_id_title_expl_artist_album.csv')

# checking the column

dr_sample_tracks.columns

# removing the not needed index column

dr_sample_tracks = dr_sample_tracks.drop(['Unnamed: 0'], axis = 1)

# checking length 

len(dr_sample_tracks)

# Getting audio features

To get the audio features I will use audio_features(tracks=[]) function.

# setting passwords

client_id = getpass.getpass('client_id?')
client_secret = getpass.getpass('client_secret?')

# connection to spotify API

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id=client_id ,
                                                           client_secret=client_secret))

# creating list of all track ids from df to iterate over to get audio features

track_ids = list(dr_sample_tracks['id'].unique())

# building a function to put the song ids in track_ids into chunks
# the audio feature function can only process 50 tracks at the time

def chunks(lst, n):
    """Yield successive n-sized chunks from lst."""
    for i in range(0, len(lst), n):
        yield lst[i:i + n]

# getting the audio features from the top100 track ids
# iterating over the top100 track list with the chunk function

dr_sample_tracks_audiofeatures = []

for chunk in list(chunks(track_ids, 50)):
    dr_sample_tracks_audiofeatures.append(sp.audio_features(chunk))

# turning the audio feature list into a df

dr_sample_tracks_audiofeatures_df = pd.DataFrame(dr_sample_tracks_audiofeatures)

# we need to unpack those dicts

dr_sample_tracks_audiofeatures_df = pd.DataFrame()

for i in range(len(dr_sample_tracks_audiofeatures)):
    try:
        dr_sample_tracks_audiofeatures_df = dr_sample_tracks_audiofeatures_df.append(pd.DataFrame(dr_sample_tracks_audiofeatures[i]))
    except:
        continue
        
dr_sample_tracks_audiofeatures_df.head()

len(dr_sample_tracks_audiofeatures_df), len(dr_sample_tracks)

# merging both dfs

dr_sample_tracks_audiofeatures_df_merged = pd.merge(dr_sample_tracks, dr_sample_tracks_audiofeatures_df, how='left', on=['id'])
dr_sample_tracks_audiofeatures_df_merged.tail()

# checking length to see how it worked

len(dr_sample_tracks_audiofeatures_df_merged)

# saving df as csv

dr_sample_tracks_audiofeatures_df_merged.to_csv('../data/dr_sample_audio_features.csv')
