==> code/audio_features/2.scraping_audio_features_top100.ipynb <==
# Getting data 3 - audio features of top100/top20 hip hop tracks

Now that I got the cleanded weekly top100 tracks for 2010-2021, I need to get the audio features of said tracks.

I got a csv with the following data:
- artist
- track title
- label
- rank
- week

<br/>

Next steps are:
- getting audio features from tracks (top100 & top20 hip hop)

# importing needed libraries

import pandas as pd
import numpy as np
import getpass
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

# importing data csv

top100 = pd.read_csv('../data/top100_tracks_V3.csv')

# Getting audio features

To get the audio features I will use audio_features(tracks=[]) function. To use this function I need the track ids.

len(top100)

top100_remove_dups = top100.sort_values('title', ascending=False)
top100_remove_dups = top100_remove_dups.drop_duplicates(subset='title', keep='first')

len(top100_remove_dups)

# setting passwords

client_id = getpass.getpass('client_id?')
client_secret = getpass.getpass('client_secret?')

# connection to spotify API

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id=client_id ,
                                                           client_secret=client_secret))

# getting the id from the track title for 1 example

track = 'drei uhr nachts'
artist = 'mark forster'

track_id = sp.search(q='artist:' + artist + ' track:' + track, type='track')

#### preparing for looping to get info from all tracks

# removing duplicates in title

track_names = list(top100_remove_dups['title'])

# removing duplicates in artist

track_artists = list(top100_remove_dups['artist'])

# zpping both lists together to dict

track_dict = dict(zip(track_names,track_artists))

# looping over dict to get all track info I need from Spotify

track_id_dict = []

for key, value in track_dict.items():
    track_id_dict.append(sp.search(q='artist:' + value + ' track:' + key, type='track'))

# extract the id for 1 example

track_id_dict[4]['tracks']['items'][0]['id']

# extract the id for 1 example

track_id_dict[4]['tracks']['items'][0]['name']

# extract the artist for 1 example

track_id_dict[4]['tracks']['items'][0]['artists'][0]['name']

# gettin the ids for all tracks

track_ids = []
track_title = []
track_explicit = []
track_artist = []

for dic in range(len(track_id_dict)):
    for track in track_id_dict[dic]['tracks']['items']:
        track_ids.append(track['id'])
        track_title.append(track['name'])
        track_explicit.append(track['explicit'])
        track_artist.append(track['artists'][0]['name'])

len(track_ids), len(track_title), len(track_artist), len(track_explicit)

# creating a df from track id, title and artist to add later to the audio feautres again

top100_trackid_title_artist_df = pd.DataFrame({'track_ids' : track_ids,
                                'track_title' : track_title,
                                'track_explicit' : track_explicit,
                                'track_artist' : track_artist }, 
                                columns=['track_ids','track_title','track_explicit','track_artist'])
len(top100_trackid_title_artist_df)

# renaming the column name track_ids to id, so its matches the id from spotify

top100_trackid_title_artist_df.columns = ['id','track_title','track_explicit','track_artist']

# saving df as csv

top100_trackid_title_artist_df.to_csv('../data/top100_trackid_title_artist.csv')

#### now getting the auio features

# getting audio features for 1 track

audio_features_test = sp.audio_features('7ue6QuZluV20e3lKVhsAy4')

# building a function to put the song ids in track_ids into chunks
# the audio feature function can only process 50 tracks at the time

def chunks(lst, n):
    """Yield successive n-sized chunks from lst."""
    for i in range(0, len(lst), n):
        yield lst[i:i + n]

# getting the audio features from the top100 track ids
# iterating over the top100 track list with the chunk function

top100_track_audiofeatures = []

for chunk in list(chunks(track_ids, 50)):
    top100_track_audiofeatures.append(sp.audio_features(chunk))

# turning the audio feature list into a df

pd.DataFrame(top100_track_audiofeatures)

# we need to unpack those dicts

top100_track_audiofeatures_df = pd.DataFrame()

for i in range(len(top100_track_audiofeatures)):
    try:
        top100_track_audiofeatures_df = top100_track_audiofeatures_df.append(pd.DataFrame(top100_track_audiofeatures[i]))
    except:
        continue
        
top100_track_audiofeatures_df.head()

len(top100_track_audiofeatures_df), len(top100_trackid_title_artist_df)

# add columns to df:
# genre
# artist name
# track title

top100_trackid_title_artist_df.columns

top100_track_audiofeatures_df_merged = pd.merge(top100_trackid_title_artist_df, top100_track_audiofeatures_df, how='left', on=['id'])
top100_track_audiofeatures_df_merged.tail()

len(top100_track_audiofeatures_df_merged)

# saving df as csv

top100_track_audiofeatures_df_merged.to_csv('../data/top100_audio_features.csv')
