==> code/audio_features/3.scraping_release_date.ipynb <==
# Getting release date for all songs

Now that I have all the tracks I need (dr sample and top 100), I also need to get the release dates in order to analyse the development over time.
<br/>The release date I will be able to get from the Spotify API via the album affiliation. As mentioned in the notebook before, I will not get a release date for all tracks as Spotify only provides the release date of the album, not on track level.
<br/>I need to anticipate NaNs.

# importing needed libraries

import pandas as pd
import getpass
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
from tqdm import tqdm

# getting data

audio_features = pd.read_csv('../data/audio_features/audio_features_all.csv')

# having a quick look at the df

audio_features.head()

# setting passwords

client_id = getpass.getpass('client_id?')
client_secret = getpass.getpass('client_secret?')

# connection to spotify API

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id=client_id ,
                                                           client_secret=client_secret))

# doing a test to get the track info for 1 example track

test = sp.track('1EFbLCDFWIIiC6IMgthYsN')
test['album']

# getting release date

test['album']['release_date']

# getting album id for merging dfs on id

test['album']['id']

# getting track info for all tracks

track_id = list(audio_features['id'])

track_album = []
track_album_release = []
track_album_id = []
track_album_name = []

for track in tqdm(track_id):
    track_album.append(sp.track(track))
    
for album in tqdm(range(len(track_album))):
    track_album_release.append(track_album[album]['album']['release_date'])
    track_album_id.append(track_album[album]['album']['id'])
    track_album_name.append(track_album[album]['album']['name'])

# checking lengths of lists for merging

len(track_album_release), len(track_album), len(track_id), len(track_album_id), len(track_album_name)

# creating a data frame 

tracks_all_release = pd.DataFrame({'track_id' : track_id,
                                'track_album_id' : track_album_id,
                                'track_album_name' : track_album_name,
                                'track_album_release' : track_album_release })

# checking if it worked

tracks_all_release.head()

# saving df to csv

tracks_all_release.to_csv('../data/tracks_all_release.csv')
